Update patch set 2
Patch Set 2: Code-Review+2

Severity: The bug impacts OPNFV and ODL integration related to SFC. SFC will not work due to not being ableto find the right port to use for service functions. It is not a Blocker for Beryllium, but more a blocker for the OPNFV and ODL SFC integration. There is not a workaround.

Testing: Manual and automated. The use case is specific to adding two flows for service function chaining that include OpenStack instantiated service functions. The flow progamming is triggered when an SFF RESTCONF API is received with br-int as the bridge. Testing was done to send the right API so that the SFC and OVSDB NetVirt code program those flows. Manual testing was done by sending the RESTCONF APIs. The automated integration tests were updated in OVSDB NetVirt to simulate the use case. In both cases the two extra flows were programmed and verified.

Impact: The patch only impacts SFC and OVSDB NetVirt in the context of the ODL OPNFV SFC integration. Without the patch the SFC will have an exception and the flows to steer traffic to and from the service functions will not be installed - effectively blocking the service function chaining. All other uses cases are not impacted.

Patch-set: 2
Label: Code-Review=+2