Update patch set 4
Patch Set 4:

> Stephen, I didn't get what your question to me was.

The approach taken here, to avoid issues with transactions, is to pass in a transaction instance to methods requiring one (instead of creating one using the data broker or expecting to be in a transaction invoker). Since this is an entirely new approach in the ovsdb codebase (AFAIK) I wondered what your thought of it.

Patch-set: 4