Update patch set 1
Patch Set 1:

> (1 comment)

Stephen et al.,

This was originally done as part of https://git.opendaylight.org/gerrit/#/c/17718/ .  I actually copied the pattern from the "options" code that Eric Multanen wrote, so I am adding him.

The ImmutableMap.copyOf(...) will throw a NullPointerException if either the key or value is null. Since HashMap API states "This implementation provides all of the optional map operations, and permits null values and the null key", this is an effort to ensure that both key and value were supplied for a given entry.  Not sure what the model looks like here, but this was an attempt to be extra careful that we weren't allowing malformed data (keyless values or valueless keys).  I don't know whether we got to an end on the question of whether "mandatory" YANG constraint prevents this from happening instead... might be worth investigating.

Patch-set: 1