Use a utility function for key-value to map conversions
YangUtils::convertKeyValueListToMap() reproduces the behaviour
expected when converting a list of YANG key values to a map: it checks
that all the keys and values are non-null, and converts them to
strings using the appropriate methods in each YANG type.

This patch can't be back-ported to Beryllium since it uses Java 8
constructs.

Change-Id: Iaeb751ca552caf6b8f5a6d5a6e506e5d2aa4e63f
Signed-off-by: Stephen Kitt <skitt@redhat.com>