Removed the NPE seen when receiving port/subnet add requests in the routemgr.
On receiving router/port/subnet northbound add requests, the NPE
obtained are resolved. Also, when adding the host interfaces,
the device ownerid obtained for the interface is not an UUID
and the device owner id is not used for associating the
interface with the subnet, so removed the configuration of the
device ownerId for the host interface.

Change-Id: I586a3f338a6da07922634f1e5ced612ec40fa026
Signed-off-by: kalaiselvik <Kalaiselvi_K@Dell.com>