Update patch set 3
Patch Set 3:

Hi Sam, This patch looks good. Alternative mechanism to handle this scenario is TCP KEEP_ALIVE, but that's not really a good way to handle it. From clustering perspective, this will help in cleaning up the connections and that will help us avoid the situation where OVSDB comes back up and uses the same port which it used before the crash, but to reduce this probability we should reduce the idle timeout and read timeout to 30 second and 180 seconds as a default. This timeout value also should be configurable, so that people can configure it according to their deployment environment.

Thoughts ?

Patch-set: 3