Create change
Patch-set: 1
Change-id: I2c720aeac44e4252091952383fe3bb4ed9bed97b
Subject: Net-virt is not checking if the master of the net-virt-application is elected at the clustering level or not. Because of this it can end up in a situation where candidates are registered but owner is not elected and all the instances gets false when they check for the owner. This patch fixes the issue, where it waits till owner is selected
Branch: refs/heads/master
Commit: 7fa749f5040b3bbc16bdd20ec002338c6f920466