Update patch set 1
Patch Set 1: Code-Review+2

Ritu, this patch looks good. I don't think there is a way to build this patch using the dependent neutron patch. We usually just hold this patch here until the neutron patch merges and then we rebase/recheck this patch to get it to verify.

So please merge the patch in neutron and we can then merge here.

Patch-set: 1
Label: Code-Review=+2