Update patch set 2
Patch Set 2:

Vishal,
(1) We want to attempt retry for controller initiated connection. You can figure out that by looking at the config data store. Library don't really have context of whether it's switch initiated connection or controller initiated connection. Moreover library will not know if this disconnect is also initiated by controller itself on user request, so we don't want to retry on that.

(2) Yup, I was thinking of that as well, will push patch to add them as a configurable parameter.

Patch-set: 2