Update patch set 3
Patch Set 3:

Actually we can know in library code if it is active or passive, OvsdbConnectionInfo.getType(). Agree, we won't know if user has actually deleted so need to retain it in connection manager to check datastore.

Question about naming: Would we be using these same tasks for configuration reconciliation or are these just for retrying connection? If it is latter, the naming will be confusing when we add full reconciliation support.

Patch-set: 3
Reviewer: Gerrit User 3276 <3276@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0