Update patch set 5
Patch Set 5:

Josh, I think the programTunnelRulesInNewNode is doing similar work to what you added in this patch. Assuming we agree on the problem. I think the problem is that when the new node comes up the handleIntefaceUpdate() will eventually program the flows for it's port to forward traffic locally and will also install flows on the dst node which I call the reverse flows so that traffic can be forwarded back to this node and to the local port. But we don't get the flows to get traffic from the new node to the existing nodes - since those reverse flows are installed as stated above and require the nodes to be existing at the time of the handleInterfaceUpdate().

So the ...newNode() method is supposed to do that. In the method it looks through the ports on the remote Nodes and then installs the reverse flows on the srcNode. So the call in handleinterfaceUpdate is already looping over all the nodes and then in newNode() is loops over the ports on the remote nodes. There is also another programTunnelRules() in newNode() programming flows on the dstbridge but I am not sure exactly what that is doing.

So it seems the newNode() method is doing similar work. But something in there is busted.

Patch-set: 5