Fixes for LibraryIT timing out
Something recently changed in the karaf builds such that
it is causing the five core karaf bundles to only go to
an installed state. The LibraryIT tests will not run unless
they see all bundles active, so the installed state bundles
cause the test to timeout.

Reworked the tests to not use the bundle states and to use what all
the other tests in OBSDB use which is to wait until the config is up.
Added these methods to the abstract test class so that it could be
shared across the five differetn library IT sets.

Also cleaned up a bunchof System.Out.println's and Assert's.

Change-Id: Iddd31eede1de1c5f3cde88bb1c3f7c96a5489051
Signed-off-by: Sam Hague <shague@redhat.com>