Update patch set 2
Patch Set 2:

(3 comments)

Hello Stephen,
What do you think of adding an ExecutorService with a threadpool that will process the transactions?
Instead of having the caller thread executing the transaction, this could be delegated to an ExecutorService, enhancing the performance of the plugin.
This might helped in a scaled setup

Patch-set: 2
Reviewer: Gerrit User 3675 <3675@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0