Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Few comments in the code. Apart from that, is there any specific requirement for using the "name" as part of the queue and not "uuid"? Is it to provider user friendly restconf interface? But user is not really allowed to add the Ucas/Mcas MAc address through config data store right? so in that case probably using the uuid is okay? If not, then i think if you can address the below comments, this patch looks good to me.

Patch-set: 1
Label: Code-Review=-1