Update patch set 1
Patch Set 1:

Vishal, did you verify that the cache gets repopulated in case of controller restart? I think it should, given that after controller restart, switch will reconnect and it will get the initial configuration from the switch, and while processing that initial configuration, it should repopulate the cache. But I think we should test it once, if this behavior is not tested.

Patch-set: 1