Update patch set 4
Patch Set 4:

> Thanks Stephen for adding the markers.
 > 
 > I am a little torn with the "passing transactions in" stuff. All of
 > the other mdsal code mostly lets the util handle the transaction
 > and allocate it itself. Seems easier than making the apps do that
 > work and keeps the exception handling with the util.
 > 
 > Could this new code keep the existing feel as much as possible?

Just to fuel in other thoughts, I think passing the transaction probably is better approach, if caller want to batch the transactions. If caller pass the transaction and utility methods submit the transaction (Except readOnlyTransaction) i don't see much point in passing the transaction in that case.

Patch-set: 4