Update patch set 4
Patch Set 4:

I think it mainly boils down to how Utils are supposed to be used. There are 2 ways to approach utils:
1. Util method should be addTunnel(nodeid, bridgename, tunnelIp, tunnelType) and it takes care of constructing identifiers and data to be written into data store and writes it.
2. Util methods should create the instance identifier and object and let applications/callers submit transactions themselves.

First one allows one to write applications that don't need to be aware of MDSAL semantics but prevents batching. Second one allows applications to batch operations but require them to be more 'MDAL aware'.

So, based on which of the two, transaction in Util will be needed or not. Thoughts?

Patch-set: 4
Reviewer: Gerrit User 3276 <3276@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0