Update patch set 4
Patch Set 4:

Along your lines Vishal, I think we need to decide on how the utils are used and then be consistent through the code. I could see the utils supporting both methods of passing or not passing the transaction. I would also like to see the utils support non-blocking transactions.

There are all different uses of the transactions throughout the code. My earlier review comments were based on trying to keep the usage consistent. I would prefer that for this review until we need to support passing in transactions to support cases like batching.

Patch-set: 4