Update patch set 3
Patch Set 3:

Is there any way to do this without having to touch all the files?

As a side but related topic, any ideas on what should be done for all the maven central info that is included in the poms? Back in Lithium there was a push to meet Maven Central requirements so we added all kinds of info in the top of the poms - url, description, licenses, developers, scm. We added that to every pom at the time. Any idea if we could have just added that to the root pom? Is there a minimum of info needed?

Patch-set: 3
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0