Update patch set 1
Patch Set 1:

(2 comments)

Just as an FYI, this is a port of a class I used in some other app. I defined the whole class as final, with static method. The DataBroker was passed as an argument. I slightly change the implementation here to better match what we already have with the MdsalUtils

Patch-set: 1
Label: Verified=0