Update patch set 1
Patch Set 1:

> I'm wondering whether it should be possible for callers to specify their own callbacks on commits, but perhaps for that kind of use-case callers should just manage everything themselves.

Exactly, this can't be used if you need to react on callback. 
Although we could tweak the request by adding a Boolean or something to let the caller choose whether or not he wants to define his own callback. In this scenario, the function would return the checkedFuture, else it would return a immediateCheckedFuture saying the transaction was correctly submitted.

What do you think?

Patch-set: 1