Update patch set 1
Patch Set 1:

Yi, yes, the files are still there but they are not included as part of the modules so they do not build. We have left the code in place for a few weeks as a reference while we verify the new netvirt code is working properly. We will likely be removing the files soon.

Patch-set: 1