Update patch set 2
Patch Set 2:

I was just wondering do we really want to submit the transaction and log on the call back methods? In general when you do these transaction, you take certain actions on the callback, specially "success". With that thought in mind i am wondering how useful this class is going to be. thoughts?

Patch-set: 2