Update patch set 2
Patch Set 2: Code-Review-1

so looking at the reference provided in the commit message it looks like you need to specifically install some library to use the epoll features.

sudo apt-get install autoconf automake libtool make tar \
                     gcc-multilib libaio-dev

Dilip, did you do any testing for openflowjava patch and check how much performance improvement did you get with and without patch? 
Looking at the code, it looks like we check if epoll is available, if not, then we fallback to NIO. But given the requirements to use epoll, this check will fail with Ubuntu 14.04 (64bit) as well, if you won't have other required libraries. So just wondering if it's worth to consume e-poll at this moment. But if we have good performance number to backup, then i think it's worth trying, but in that case as well, i think we should provide a user configuration to enable/disable it.

Patch-set: 2