Update patch set 2
Patch Set 2:

Anil, you're right, some calls would react onSuccess or onFailure, but some other doesn't care and could just be made async to avoid blocking the caller thread. So this patch is taking care of the later. If someone wants to define its own callback, then not using this class is the right way to do it.

I made this class because I've said in couple of review that databroker transactions could be made async.

Thinking about it maybe having a wiki page explaining the how to vs this patch could make sense.

Just shouting ideas here.

Patch-set: 2