BUG 5885 - OVSDB plugin failure to update passive ovsdb nodes
Updates the ovsdb southbound so that configuration to ovsdb
nodes which have connected to the plugin passively can be
updated (e.g. such as qos or queue entries) without requiring
connection-info to be supplied in the configuration.
Similar to what already works when adding bridges or
termination points to passively connected ovsdb nodes.

Change-Id: If0c819d8e4c741e01091d272043bc8a902806b87
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>