BUG 5885 - OVSDB plugin failure to update passive ovsdb nodes
Updates the ovsdb southbound so that configuration to ovsdb
nodes which have connected to the plugin passively can be
updated (e.g. such as qos or queue entries) without requiring
connection-info to be supplied in the configuration.
Similar to what already works when adding bridges or
termination points to passively connected ovsdb nodes.

The onDataTreeChanged() method executes the following
sequence of operations on changes to the Ovsdb data tree:
- connect(changes)
- updateConnections(changes)
- updateData(changes)
- disconnect(changes)

When an OVSDB node has connected passively, or more
generally, when the changes to the OVSDB node do not
include 'connection-info', it is not necessary to
attempt to execute connection related operations for
the specified host.  This patch  checks for the absence
of the connection-info and skips trying to connect or
update a connection in the connect() and updateConnections()
methods.

In the updateData() method, an extra condition is added
to the connectionInstancesFromChanges() method to look up
the OvsdbConnectionInstance by node instance identifier
if the connection-info has not been supplied for the
ovsdb node.

v2 - fix whitespace
v3 - add more details on the change to the commit message
     fixed indentation of log message
v4 - rebase

Change-Id: If0c819d8e4c741e01091d272043bc8a902806b87
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>