Bug 5721 - br-int not created in clustered setup
When the OVS node is hard-reset in PASSIVE mode, it takes
up to 3 minutes for the controler to detect the connection is gone.
During this time, new connection request from the same node arrives.
The latter connection request causes
CandidateAlreadyRegisteredException because the former connection
has not been cleaned up properly.

This patch fixes the issue by:
 - shorten the the netty channel inactivity timeout to 15 sec.
 - When the IdleStateEvent is triggered due to channel inactivity,
   the controller pings the connection and synchronously checks
   if the ping is sent successfully. If not then we know the connection
   is definitively down and close it.

Change-Id: I8bbce8a81b9850704278220b4b75d2ab21cb34c6
Signed-off-by: vinh.nguyen <vinh.nguyen@hcl.com>