Bug 5721 - br-int not created in clustered setup
When the OVS node is hard-reset in PASSIVE mode, the 'disconnected'
event might not properly received before the 'connected' event.
As a result the latter 'connected' event causes
CandidateAlreadyRegisteredException because the former connection
has not been cleaned up properly due to the fact that
'disconnected' event is not received.

This patch fixes the existing logic to check and disconnect existing
client PASSIVE connection in OvsdbConnectionManager::connected method.
Only the remote IP is matched, The remote port is ignored since the latter
is changed in each connection attempt.

Change-Id: I8bbce8a81b9850704278220b4b75d2ab21cb34c6
Signed-off-by: vinh.nguyen <vinh.nguyen@hcl.com>