Update patch set 1
Patch Set 1:

I think we need to revisit ReadtimeoutHandler now that reconiliation logic is in. What am observing is [with this fix] that after 60 secs we try sending out an echo which results in disconnect in about 18 seconds which then triggers Reconciliation logic. Could it be that with connection reconciliation we don't need ReadTimeoutHandler, or that ReadTimeoutHandler isn't working as expected?

Patch-set: 1