Update patch set 1
Patch Set 1:

Not sure if this is the proper fix but this worked in my setup. Maybe need to add this to odl-ovsdb-library? There is also room for improvement in how both plugins are using odl-ovsdb-library, there seems to be some redundancy in there.

Patch-set: 1
Reviewer: Gerrit User 3276 <3276@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0