Update patch set 1
Patch Set 1:

Hi Vinh, I want to suggest an alternate mechanism to handle this situation. We should handle this at the library level. So whenever we get connection from the same machine, we should disconnect the existing connection and sent the disconnect notification up and then process the next connection. That will take care of the cleanup at the southbound plugin. The reason behind this thought is that we should not do the connection handling at the plugin level, because if we start handling these scenario, it can get really very complex to handle it in 3-node clustering environment.

Patch-set: 1