Update patch set 1
Patch Set 1: Code-Review-1

I also agree I think it would be better handled in the library, but we need to make sure the handling stays simple. Meaning, we can have simple heuristics to determine this is the same connection so we can disconnect it. What is the node supports multiple schemas and connects with both open_vswtich and hw_vtep - they would both come from the same ip, not sure about port.

Patch-set: 1
Label: Code-Review=-1