Update patch set 1
Patch Set 1: Code-Review-1

Hi Sam/Anil

Do you think we can determine that it is the same connection if it comes from the same node and have the same list of supported db schemas?

I tried to use netty's channel handler to listen to channelInactive/channelClosed events but the events don't happen when the connection is hard-reset. This seem to be a common issue that can only be addressed with connection pings

Patch-set: 1
Label: Code-Review=-1