Update patch set 1
Patch Set 1:

I'll push in another patch that fixes the problem using org.opendaylight.ovsdb.lib.jsonrpc.ExceptionHander and netty'sIdleStateHandler to detect the netty channel inactivity.  Currently the OVS node sends "echo" to the controller every 5 seconds. The readerIdleTimeSeconds of the IdleStateHandler will be set to 15 seconds so that it raises an IdleStateEvent if there are 3 echo misses. When the ExceptionHander receives the IdleStateEvent event, it tries to ping to switch one more time. If the ping still fails the controller closes the channel/connection.

This is consistent with the openflowplugin's switch connection handling

Patch-set: 1