Bug 5951 - Termination point config reconciliation
Add termination point configuration reconciliation logic
to southbound plugin.

When the OVS switch connects to the ODL, the list of all bridge
and termination point configurations in the config datastore
is obtained. We then listens for any new bridge creations in the
operational DS. If the newly created bridge is in the above described
list, termination point reconciliation is triggered for that bridge.

This is the solution for reconciliation scenarios #5 and #7 in:
https://wiki.opendaylight.org/view/OVSDB_Integration:OVSDB_SB_Reconciliation

Change-Id: Ie0fc8bcbd12a5c1dd763a580db7097a4afd7e01a
Signed-off-by: Vinh Nguyen <vinh.nguyen@hcl.com>