Update patch set 4
Patch Set 4:

Ananthi, In the logging improvement i am looking at the following improvement (1) We should dump the whole Ovsdb/Bridge/TErminationPOint augmentation data in the INFO logs, we should just extract the relevant information and log them. (2) We should all more logs that can help us in debugging the plugin (3) Set the log level appropriately. The result should be that when plugin start we see very trimmed logs but those logs should be able to established that plugin booted up fine. And looking at the debug log it should provide us all the information so that we don't have to re-run the plugin/usecase to find out the root cause.

Patch-set: 4