Update patch set 4
Patch Set 4: Code-Review-1

Hi Vinh, Current implementation actually wait for bridge configuration to be done before it triggers the termination point configuration. We don't have to follow this order. We can send both bridge configuration and termination configuration in the same transact. Plugin will send all the transact in the order (bridge, termination point) and ovs will process them in that order. Assume this scenario like, user is creating bridge and termination point in the same restconf call. In this scenario plugin create all the transact rpc messages and dump in single transaction. We can use the same approach for reconciliation and that will be cleaner approach. Please drop me a mail if you want to discuss it further.

Patch-set: 4
Label: Code-Review=-1