Update patch set 1
Patch Set 1: Code-Review-1

i don't think we want to have this improvement in stable/beryllium branch, because we are deprecating the yang model that we introduced in the stable/beryllium in this patch. So if we cherry-pick it to stable/beryllium, i don't see why we want to deprecate stuff and just remove the deprecated models.

Patch-set: 1
Reviewer: Gerrit User 65 <65@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1