BUG 6130 - process only qos and queue creates and updates
Fixes issue where all qos or queue entries - including ones
that had not changed - for an ovsdb node were getting ovsdb
transactions sent to the ovs node when a qos or queue was
created or updated.

This patch only creates ovsdb transactions for qos or queue
entries that have been created or updated.

Change-Id: Ie9eef13deae768dbdc16d6ab2a35e353c1aba6bf
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>