Bug 6185 - southbound system tests failing when run with all other compatible OpenDaylight features.
This patch fixes following issues related to library
connection handling
*.* If any non-ovsdb device (e.g. openflow) sends a
connection request to ovsdb, library does not throw
error message, rather it holds the channel and wait
for timeout.
*.* In the above case, it also publishes the connect/disconnect
notification up to the plugin.
*.* HwvtepConnectionInstance and OvsdbConnectionInstance
is implementing OvsdbClient interface from library, that
is not required. Echo service should be done by the library
and not by the plugin logic code.
*.* RPC Future objects were not cleaned by library,
and that's leaking the memory.
*.* In case of no echo response, library dosn't fail
the respective rpc future object.

Change-Id: If7b26594b057465d0c08638857bdcb76857705e1
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>