Update patch set 2
Patch Set 2:

The OVSDB Plugin registration is registering for a all the data, including statistics, status and version so the OVSDB is "pushing" All the data to ODL every 1-3 seconds. To be able to scale and not receive all the data every 1-3 we remove those fields from the registration so we "loose" a subset of data that initially was there.
To scale and still keep that data for those fields we need to switch the method of fetching this data from the original "register+get notification" to "poll", e.g. for each one of those fields ODL controller needs to explicitly request this data every X interval and fill the "gap" in the model with the result of the query.
For example if I remove the status from the original registration because with it I get all the OVSDB data every 1-3 seconds, I should start polling only the status explicitly using the OVSDB query mechanism, e.g. the other way (or the traditional way) of submitting a query->receiving the result->update the model-> do that every X interval that proves to be efficient.

Patch-set: 2