Fix clear bug related to "num" in JSON Node [Guava instead JDK Optional]
Before this change, getConstraints() would call populateEnum()  which
would ... do nothing, ultimately; this was pointless, as it is currently
implemented.

I'm assuming here that the code was meant to do what I turned it into,
but I don't understand enough of this code to be certain.  I've also NOT
tested this in any way.  It does kind of seem reasonable though, looking
at what the lines before in getConstraints do.

An alternative would be to just remove those populateEnum methods - if
they do nothing, and there is no real bug here, then that was just dead
code.  More likely, I suspect it is a real bug that hasn't been found
yet.

I'll try to add FindBugs error detection for such cases in the future -
UC_USELESS_OBJECT is exactly for finding these kind of programming
errors, AFAIK.

Change-Id: I587e053b5145320b600e2db0109118d799da8822
Signed-off-by: Michael Vorburger <vorburger@redhat.com>