Update patch set 1
Patch Set 1: Code-Review-1

In my opinion, we should not put utility methods in OVSDB project. These project should be part of the project that's consuming the ovsdb. Reason is if project start putting these utility methods, they are putting another dependency on the project level and all the consumer project will start pushing these changes in the ovsdb project. So these utils method will act like unannounced APIs and if any project modify these utils method, which breaks other projects, this will be additional overhead for the OVSDB project to manage and in my opinion that's not a good idea.

Patch-set: 1