Fix connection reconciliation regression
After enabling the checkstyle, we start catching the
specific exceptions rather then catching all the exception
using Exception. During the retry, connection task encounters
ConnectionException, but nobody catches that exception and
it kills the reconciliation thread, so it does not retry
the connection after first attempt. This patch fixes this
issue and handles the ConnectionException properly.

Change-Id: I0e1f71ea6fa550df484ade9036f7dfe6c9dfa57e
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>