BUG 6467: Add cli flag to use virtual environment
This change is a followup to patch:
 https://git.opendaylight.org/gerrit/#/c/44022/

Add cli flag for virtual environment

Docker-compose is installed in a virtual environment on Jenkins.

Add a cli flag to enable switching to a virtual environment, if needed.

Netvirt IT requires docker-compose, so the netvirt IT jenkins
script will set this flag: -Ddocker.vEnvWs=$WORKSPACE

Change-Id: I989d071ef2303f4b4813c3145435b2ddb785a228
Signed-off-by: Victor Pickard <vpickard@redhat.com>