Update patch set 2
Patch Set 2: Code-Review+1

Hi Anil, thank you for your comment!

I understand performance is always dependent on where you are running the controller.

However, my point is that I think if the number hits 50 or 100, it means that the speed of which the plugin process update notification is slower than the speed of update notification coming from OVS for a while. So the situation means that there is a risk that users hit critical issues such as bug 5376 in which the OVSDB southbound plugin took several minutes to process a update notification so that NetVirt failed to install flow entries quickly so that VM cannot reach DHCP server when it was needed. To avoid critical issues, the plugin should process update notifications as quickly as the plugin receives update notification for most time. So I would say that the number of command pending should be less than 1. Of course, it's ok the speed is slower than the speed of update notification for a while, but if such kind of situation continues for a long time, it's not ok, so we should warn users.

So yeah, if I am running the controller in 1 CPU/2gig ram, we probably will hit 50 very easily, it means we will easily face issues such as the bug 5376. Therefore, I think users should be warned in such a situation. Otherwise, users, which use 1 CPU / 2gig ram, probably get confused when they face issues, since ODL does not warn and does not give any hints why the issue occurred.

The OVSDB Southbound plugin is responsible for updating topology-operational DS based on update notification coming from OVS. Therefore, I think when there is a problem on updating topology-operational DS, the OVSDB Southbound plugin should warn users. Without this patch, how can users know the reason why topology-operational DS is not updated after they connects OVS into ODL? (debug logs are for developers, but not users, so I don't expect users enable DEBUG log.;)

I don't think the warning message would be a false positive. The ovsdb plugin is who processes update notification. And, the warning message which the patch tries to output is about the delay on processing update notifications and update topology-operational DS. I think it will be always true positive. Please note that warning message does not say that there is a performance issue in the plugin. (I'm feeling I don't correctly understand what you are trying to say here... So my response might not make sense.)


I think it's good to print some status (queue size and any other stats) in debug level, but it is for developers. So that is different from what I'm trying to achieve here.

Though, it depends on many factors, so it should be a configurable parameter in the future.

Patch-set: 2
Reviewer: Gerrit User 396 <396@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1