Update patch set 2
Patch Set 2: Code-Review-1

> Hi Anil, thank you for your comment!
 > 
 > I understand performance is always dependent on where you are
 > running the controller.
 > 
 > However, my point is that I think if the number hits 50 or 100, it
 > means that the speed of which the plugin process update
 > notification is slower than the speed of update notification coming
 > from OVS for a while. So the situation means that there is a risk
 > that users hit critical issues such as bug 5376 in which the OVSDB
 > southbound plugin took several minutes to process a update
 > notification so that NetVirt failed to install flow entries quickly
 > so that VM cannot reach DHCP server when it was needed. To avoid
 > critical issues, the plugin should process update notifications as
 > quickly as the plugin receives update notification for most time.
 > So I would say that the number of command pending should be less
 > than 1.
Thanks not a practical number, here we are talking about one plugin that is managing multiple device (ideally more then 100+) in parallel, so you will never achieve this number even if you run your plugin at high scale server. Even if you spawn a dedicated thread for each device, that won't achieve the same because you are writing in the same tree in the topology. That's the reason we are using the queuing at first place.
 Of course, it's ok the speed is slower than the speed of
 > update notification for a while, but if such kind of situation
 > continues for a long time, it's not ok, so we should warn users.
what do you expect user to do when this warning appears? I am not sure user can take any actions. This warning is more useful for the developer or the person who can debug the system or suggest some solution on how it can be prevented. What solution are you planning to propose when you this warning appear? In my opinion, it's something that a developer/support can use to diagnose the system, so it's very much related to the developer rather than end user. With that perspective, only throwing warning won't help, you will have to provide more debug info to the developer/support person to figure out the kind of load plugin is facing, so that can help them to diagnose the issue. 
 > 
 > So yeah, if I am running the controller in 1 CPU/2gig ram, we
 > probably will hit 50 very easily, it means we will easily face
 > issues such as the bug 5376. Therefore, I think users should be
 > warned in such a situation.
again the same question, what user is supposed to do with that warning?
 > Otherwise, users, which use 1 CPU /
 > 2gig ram, probably get confused when they face issues, since ODL
 > does not warn and does not give any hints why the issue occurred.
And that's where I personally think spitting transaction queue stats will be helpful for them rather then warning them about the long queue size, which they can't really do anything about it.
 > 
 > The OVSDB Southbound plugin is responsible for updating
 > topology-operational DS based on update notification coming from
 > OVS. Therefore, I think when there is a problem on updating
 > topology-operational DS, the OVSDB Southbound plugin should warn
 > users.
What if it's some other project that is doing heavy transaction and causing the delay in executing the ovsdb plugin transaction, although there are very few devices connected to the plugin? If plugin warns, it's a false positive from plugin perspective, because it's a bottleneck at the mdsal/datastore level. You are basically asking user/developer to look at the wrong place to solve the problem.
 > Without this patch, how can users know the reason why
 > topology-operational DS is not updated after they connects OVS    >into

...and that reason is that plugin is slow in processing the transaction or it's md-sal/datastore is heavily used by other project, that's causing the queue buildup?  or it;s because read operation is slow on the data store that's causing the issue? what should user think? Let's see user assume one of the reason and then what do you expect user to do?
 
 > ODL? (debug logs are for developers, but not users, so I don't
 > expect users enable DEBUG log.;)
I think this warning is useless for users, until and unless we provide them some config/tool/suggestion to solve the problem.
 > 
 > I don't think the warning message would be a false positive. 
Please see my above comment, when it can be false positive.
 > The
 > ovsdb plugin is who processes update notification. And, the warning
 > message which the patch tries to output is about the delay on
 > processing update notifications and update topology-operational DS.
 > I think it will be always true positive. Please note that warning
 > message does not say that there is a performance issue in the
 > plugin.
Putting that as a warning probably will confuse user. threshold=50 might be okay if you connect 10 device, but this number might not be appropriate for 20 device and when you connect 20 devices it will keep spitting the warning. 
 (I'm feeling I don't correctly understand what you are
 > trying to say here... So my response might not make sense.)
 > 
 > 
 > I think it's good to print some status (queue size and any other
 > stats) in debug level, but it is for developers. So that is
 > different from what I'm trying to achieve here.
Till user can do anything about the warning, this warning will be useful for the developers only.
 > 
 > Though, it depends on many factors, so it should be a configurable
 > parameter in the future.
As you said, it depends on many factor, so coming up with any specific threshold value is not possible.

In my opinion, spitting the warning message can cause false positive from ovsdb plugin perspective, so i would personally don't like this approach that can possible throwing warning message in log forever. But i would like to add other committers for the review and if they think otherwise, i would be happy to merge this patch.

Patch-set: 2
Reviewer: Gerrit User 65 <65@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1