Fixed inappropriate WARN message.
The plugin output the following inappropriate WARN message sometimes.
    Found non-topological node Optional.absent() on path {}
For example, this message was output when an ODL node got connected with
OVS, but the ODL node was not the owner of the OVS.

This patch fixed the inappropriate message, and changed the log level
from WARN to DEBUG, since it is a debug information.

Other change:
  Modified the code which used Optional<> in an ineffective way.

Change-Id: I714a10b509e9a88676e910ec9fba717800c52272
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>