Update patch set 2
Patch Set 2: Code-Review-1

Hi Hideyuki, 
We need a bug for this patch as we are in RC2 time period. We will have to hold it as well kill we release the carbon, because it's not really a blocking bug. I cherry-picked the patch to master and stable/beryllium and master and i will merge the patches.

Patch-set: 2
Label: Code-Review=-1